Updated the right table names onto the income modal
Incorporated the right table names onto the income modal to be able to call and fetch to them. Still need to test out the connection in order to get a respond and to POST the income data onto the database AWS.

Also the income types were updated to the correct table names.