Light and Dark Mode Working Now
The light and dark mode feature is now working. Still needs to be implemented to every page and component. So far ScreenWrapper and Welcome Screen are done. The react hook of theme is working with minor tweaks to code and styling. Remember to use const useTheme(); and import the correct theme Function.

Named: import { useTheme } from '@/context/ThemeContext';

from ThemeContext.tsx file